backend/requirements.txt
django
djangorestframework
django-cors-headers
django-debug-toolbar
psycopg2
gunicorn